sched/fair: Don't let tasks slip away from gold to silver cluster
We often see a BIG/medium task being migrated from gold cluster to
silver cluster during load balance when the gold CPU has 2 tasks
including this BIG task. For example, BIG task wakes up a binder
and binder preempts and running. At this time, the silver CPU pull
the BIG task since the overall utilization is more. The gold CPU
becomes idle immediately and may not be able to pull the medium
task. It can also happen when a pinned kworker runs on the gold
CPU.

Change-Id: I7d8caad6978bfa93aa3cdd772400e807b9f15e3d
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>